Avoid stack overflow in R2Stretcher::study.
The `study` method calls `alloca` in a loop, which causes a large amount of data to be allocated on the stack - proportional to the provided input. (`alloca` does not follow C++'s RAII rules; the allocated memory stays on the stack until the function exits, rather than the enclosing scope.) With a large enough input, this can cause a stack overflow.

This commit avoids the problem by only calling `alloca` at most once and re-using the buffer if necessary.